Copy imports from source files
The interface file has to import the same packages
because the types have to be the same.
If you have multiple packages with the same name,
goimports might not import the correct one.
Copying over the original import statements from the source file
fixes this.
If we copy over unneeded packages, goimports will remove them.

Fixes #6